done some refactor
