+++ public/modules/createLinks.js
